configs: updated newport defconfig
- removed non ThunderX architectures
- removed non ThunderX host interface drivers
- added majority of USB/PCI device drivers

The resulting kernel dropped from 17MB to 11MB (faster boot) and adds
modules for the majority of the peripheral devices which can be
added to the USB/PCI bus. The on-board NIC and storage device drivers
were left static to allow mounting a root filesystem over those devices
without needing an initial ramdisk.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>