can: mcp251x: configure the interrupt type for level triggered
The MCP2515 has a level triggered interrupt that will not be de-asserted
until all interrupt conditions are handled.

Additionally, configuring it as a falling edge triggered interrupt will
cause the ISR to be called immediately within the open function that
registers it and if for some reason the open fails it will de-register
the interrupt which will deadlock if the isr has fired and is currently
blocking on the mutex.

Signed-off-by: Tim Harvey <tharvey@gateworks.com>