configs: build DP83867 phy driver as static
CONFIG_DP83867_PHY should be set as static to ensure its driver claims the
RGMII and SGMII phy on various Newport boards. This addresses an issue using
the generic phy driver where 100mbit POE injectors connected to gig ports
would fail to auto negotiate correctly and subsequently fail to pass traffic.

The generic phy driver works as is in this case for the QSGMII interfaces on
GW64xx boards with the VSC8574 phy.

Signed-off-by: Robert Jones <rjones@gateworks.com>